Support clones from Azure DevOps
In general, go-git can't clone from Azure DevOps, because the latter
requires the capabilities multi_ack and multi_ack_detailed, which aren't
implemented. However, there's now a workaround, which boils down to
this: pretend, for the initial clone, that those capabilities _are_
supported, and expect them not to be used.

(See https://github.com/go-git/go-git/pull/613 for more on this
workaround.)

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>